feature(app): interfaces improvements
