+++ packages/app/src/api/repositories_OLD/dashboardRepository/dashboardRepository.api.types.ts
@@ -11,17 +11,6 @@ export interface TilesUpdatePositionInterface {
   data: TileListInterface;
 }
 
-export interface TextTileFormInterface {
-  type: string;
-  text_title: string;
-  text_description: string;
-}
-
-export interface VideoTileFormInterface {
-  type: string;
-  youtube_url: string;
-}
-
 export interface CreateTileDataRequest {
   type?: TileTypeEnum;
   content?: ContentInterface;
