+++ packages/app/src/core/interfaces/common/coordination.interface.ts
