+++ packages/app/src/core/interfaces/common/magicLink.interface.ts
