+++ packages/app/src/core/interfaces/common/message.interface.ts
