+++ packages/app/src/core/interfaces/common/plugin.interface.ts
