+++ packages/app/src/core/interfaces/common/scanner.interface.ts
