+++ packages/app/src/core/interfaces/common/stageModePopupInfo.interface.ts
