+++ packages/app/src/core/interfaces/common/stageModePopupOptions.interface.ts
