+++ packages/app/src/core/interfaces/forms/dashboard.interface.ts
@@ -1,0 +1,10 @@
+export interface TextTileFormInterface {
+  type: string;
+  text_title: string;
+  text_description: string;
+}
+
+export interface VideoTileFormInterface {
+  type: string;
+  youtube_url: string;
+}
