+++ packages/app/src/core/interfaces/forms/emoji.interface.ts
@@ -1,0 +1,4 @@
+export interface EmojiFormInterface {
+  file: File;
+  name: string;
+}
