+++ packages/app/src/core/interfaces/forms/events.interface.ts
@@ -1,0 +1,9 @@
+export interface EventFormInterface {
+  start: Date;
+  end: Date;
+  title: string;
+  hosted_by: string;
+  web_link: string | null;
+  description: string;
+  image?: File;
+}
