+++ packages/app/src/core/interfaces/forms/signUp.interface.ts
@@ -1,0 +1,7 @@
+export interface SignUpFormInterface {
+  avatar?: File;
+  avatarHash?: string;
+  name?: string;
+  profileLink?: string;
+  bio?: string;
+}
