+++ packages/app/src/core/interfaces/index.ts
@@ -1,13 +1,2 @@
-export * from './route.interface';
-export * from './web3.interface';
-export * from './magicLink.interface';
-export * from './scanner.interface';
-export * from './unityPosition.interface';
-export * from './unityApi.interface';
-export * from './stageModePopupInfo.interface';
-export * from './stageModePopupOptions.interface';
-export * from './message.interface';
-export * from './coordination.interface';
-export * from './tokenRulesColumnHeader.interface';
-export * from './emoji.interface';
-export * from './plugin.interface';
+export * from './common';
+export * from './forms';
