+++ packages/app/src/scenes/collaboration/pages/DashboardPage/components/templates/TileForm/components/VideoTileForm/VideoTileForm.tsx
@@ -5,7 +5,7 @@ import {toast} from 'react-toastify';
 import {useTranslation} from 'react-i18next';
 import {Button, Input, Loader} from '@momentum-xyz/ui-kit';
 
-import {VideoTileFormInterface} from 'api';
+import {VideoTileFormInterface} from 'core/interfaces';
 import {TOAST_COMMON_OPTIONS, ToastContent} from 'ui-kit';
 import {YOUTUBE_URL_PLACEHOLDER} from 'core/constants';
 import {TileInterface} from 'core/models';
